Refactor rename file to camelCase
